\name{BIdata}
\alias{BIdata}
\alias{ubi.data}
\alias{pbi.data}
\alias{pbi.class}
\alias{ubi.class}
\alias{pbi.id}
\docType{data}
\title{
Paired and unpaired metabolic data sets.
}
\description{
Paired (pbi.data) and unpaired metabolic data set (ubi.data) including classlabels (pbi.class, ubi.class). 
For pbi.data an additional vector of the ids is available (pbi.id). 
}
\usage{data(BIdata)}
\format{
\tabular{lll}{
Name of variable  \tab  Description     \tab  Type \cr
pbi.data  \tab  Paired metabolic dataset (60 samples, 60 features)   \tab  Matrix \cr
pbi.class  \tab  Classlabels for the paired metabolic dataset   \tab  Vector \cr
pbi.id  \tab  IDs for the paired metabolic dataset   \tab  Vector \cr
ubi.data  \tab  Unpaired metabolic dataset (52 samples, 60 features)   \tab  Matrix \cr
ubi.class  \tab  Classlabels for the unpaired metabolic dataset   \tab  Vector 
}
}
\details{
}
\source{
}
\references{
}
\examples{
data(BIdata)
}
\keyword{datasets}
