\name{BiomarkeR-internal}
\title{Internal BiomarkeR objects}
\alias{.calcPercentageChange}
\alias{.calculateRatios}
\alias{.createGraph}
\alias{.getDA}
\alias{.getEdgelist}
\alias{.getGraph}
\alias{.getSignificantValues}
\alias{.getTP2}
\alias{.plotScores}
\description{Internal BiomarkeR objects.}
\details{These are not to be called by the user.}
\keyword{internal}